change files to ts
